refactor(core): split models.py into a models/ package
Splits the 496-line apps/core/models.py by concern: mixins (abstract
bases), constraints (helpers), license (the one concrete model), and
references (the RecordReference graph). __init__.py re-exports every
public name so the 55 existing import sites keep working unchanged.
